Close modal with browser Back button
Uses the html5 dom for browser history to listen for a popstate event and close the modal if a modal is open
Requires pushing into history when the modal is open so that the popstate event doesn't leave the page
There's a bug now where the history fills up with modal navigations, meaning the back button would have to be pressed multiple times to leave the page after a modal has been opened 1 or more times. A solution is proposed in comments, but fails in testing.